the only thing tests still work is that I didn't switch to the new Prover implementation yet :P
