+++ aitools/proofs/exceptions.py
@@ -1,0 +1,2 @@
+class UnsafeOperationException(Exception):
+    pass
