--- aitools/proofs/proof.py
@@ -1,31 +1,0 @@
-from __future__ import annotations
-
-from dataclasses import dataclass
-from typing import Iterable
-
-from aitools.logic import Substitution, Expression
-
-
-class Prover:
-    def __call__(self, formula: Expression, _kb=None, _truth: bool = True,
-                 _previous_substitution: Substitution = None) -> Iterable[Proof]:
-        raise NotImplementedError
-
-
-@dataclass(frozen=True)
-class Proof:
-    inference_rule: Prover
-    conclusion: Expression
-    substitution: Substitution
-    premises: Iterable[Proof] = ()
-
-
-class ProofSet:
-    def __init__(self, proofs: Iterable[Proof]):
-        self._proofs = proofs
-
-    def __iter__(self):
-        return self._proofs.__iter__()
-
-    def __bool__(self):
-        return any(self._proofs)
